pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "fastapi[all]>=0.115.12",
    "litestar[jwt]>=2.15.2",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.8.1",
    "python-jose>=3.4.0",
    "sqlalchemy>=2.0.40",
]
